Fix reverse ordering of same name headers
When received multiple Set-Cookie headers in same response, after parseHeaderOutput the cookies order is reversed. 
Can break in situations where order of cookies is important because later in guzzlehttp/guzzle setCookie is changing the content of the same-name cookie. Therefore instead of eventually ending in cookieJar with the newest cookie value, we are left with the oldest.
This for example has the effect of breaking secure logins in many servers.